refactor(kebab): refactor file structure
